Finishing Upload Data Pinjaman Initial Commit Pencairan Pinjaman

+++ inkubator-hrm/src/main/java/com/inkubator/hrm/dao/LoanDao.java
@@ -28,5 +28,5 @@ public interface LoanDao extends IDAO<Loan> {
 
     public List<Loan> getAllDataByEmpDataId(Long empDataId);
     
-    public Integer getCurrentMaxId();
+    public Long getCurrentMaxId();
 }
